requirements.txt
requests==2.26.0
twilio==6.63.2
Flask==2.0.1
python-dotenv==0.19.0
spotipy==2.19.0
